+++ CHANGES.md
@@ -8,6 +8,7 @@ $ cd gsa && git log
 
 ## gsa 8.0 unreleased
 
+ * Don't crash if start or end date for performance page are invalid #1237
  * Change order of options in target dialog #1233
  * Don't limit the input field lengths anymore #1232
  * Convert first filter keyword values less then one to one #1228